+++ src/main/kotlin/com/github/xeonkryptos/integration/gitlab/api/model/GitlabProject.kt
@@ -1,0 +1,7 @@
+package com.github.xeonkryptos.integration.gitlab.api.model
+
+/**
+ * @author Xeonkryptos
+ * @since 17.09.2020
+ */
+data class GitlabProject(val name: String, val gitlabHost: String)