+++ src/main/kotlin/com/github/xeonkryptos/integration/gitlab/service/GitlabData.kt
@@ -1,0 +1,3 @@
+package com.github.xeonkryptos.integration.gitlab.service
+
+data class GitlabData(val gitlabHosts: MutableSet<String>)