+++ src/main/kotlin/com/github/xeonkryptos/integration/gitlab/ui/cloneDialog/GitlabCloneDialogExtensionComponent.kt
@@ -1,36 +1,25 @@
 package com.github.xeonkryptos.integration.gitlab.ui.cloneDialog
 
+import com.github.xeonkryptos.integration.gitlab.api.GitlabProjectApi
 import com.github.xeonkryptos.integration.gitlab.bundle.GitlabBundle
-import com.github.xeonkryptos.integration.gitlab.ui.component.TreeWithSearchComponent
+import com.github.xeonkryptos.integration.gitlab.service.GitlabDataService
 import com.github.xeonkryptos.integration.gitlab.util.GitlabNotifications
 import com.github.xeonkryptos.integration.gitlab.util.GitlabUtil
-import com.intellij.dvcs.repo.ClonePathProvider
 import com.intellij.dvcs.ui.CloneDvcsValidationUtils
-import com.intellij.dvcs.ui.DvcsBundle
-import com.intellij.dvcs.ui.SelectChildTextFieldWithBrowseButton
-import com.intellij.openapi.fileChooser.FileChooserDescriptorFactory
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.ui.ValidationInfo
 import com.intellij.openapi.vcs.CheckoutProvider
 import com.intellij.openapi.vcs.ui.cloneDialog.VcsCloneDialogExtensionComponent
 import com.intellij.openapi.vfs.LocalFileSystem
 import com.intellij.openapi.wm.IdeFocusManager
-import com.intellij.ui.ScrollPaneFactory
-import com.intellij.ui.SearchTextField
 import com.intellij.ui.components.panels.Wrapper
-import com.intellij.ui.layout.LCFlags
-import com.intellij.ui.layout.panel
-import com.intellij.ui.treeStructure.Tree
 import com.intellij.util.containers.ContainerUtil
 import com.intellij.util.ui.JBEmptyBorder
 import com.intellij.util.ui.UIUtil
 import git4idea.checkout.GitCheckoutProvider
 import git4idea.commands.Git
-import git4idea.remote.GitRememberedInputs
 import java.nio.file.Paths
 import javax.swing.JComponent
-import javax.swing.tree.DefaultMutableTreeNode
-import javax.swing.tree.DefaultTreeModel
 
 /**
  * @author Xeonkryptos
@@ -43,43 +32,38 @@ class GitlabCloneDialogExtensionComponent(private val project: Project) : VcsClo
         private val LOG = GitlabUtil.LOG
     }
 
-    private val wrapper: Wrapper = Wrapper()
-    private val searchField: SearchTextField
-    private val tree: Tree
-    private val directoryField = SelectChildTextFieldWithBrowseButton(ClonePathProvider.defaultParentDirectoryPath(project, GitRememberedInputs.getInstance())).apply {
-        val fcd = FileChooserDescriptorFactory.createSingleFolderDescriptor()
-        fcd.isShowFileSystemRoots = true
-        fcd.isHideIgnored = false
-        addBrowseFolderListener(DvcsBundle.getString("clone.destination.directory.browser.title"), DvcsBundle.getString("clone.destination.directory.browser.description"), project, fcd)
+    private val wrapper: Wrapper = Wrapper().apply { border = JBEmptyBorder(UIUtil.PANEL_REGULAR_INSETS) }
+
+    private val cloneRepositoryUI: CloneRepositoryUI = CloneRepositoryUI(project).apply {
+        addClonePathListener { newClonePath ->
+            clonePath = newClonePath
+            dialogStateListener.onOkActionEnabled(clonePath != null)
+        }
     }
 
+    private val gitlabProjectApi: GitlabProjectApi = GitlabProjectApi(project)
+    private val gitlabDataService = GitlabDataService.getInstance(project)
+
+    private var clonePath: String? = null
+
     init {
-        val treeRoot = DefaultMutableTreeNode()
+        val tokenLoginUI = TokenLoginUI(project) {
+            val projectList = gitlabProjectApi.retrieveProjectList()
+            cloneRepositoryUI.updateProjectList(projectList)
 
-        // TODO: Dummy test data. Remove later
-        val xeonkryptosNode = DefaultMutableTreeNode("Xeonkryptos")
-        xeonkryptosNode.add(DefaultMutableTreeNode("eclipse-project-creator"))
-        treeRoot.add(xeonkryptosNode)
-        treeRoot.add(DefaultMutableTreeNode("funartic"))
-
-        val treeModel = DefaultTreeModel(treeRoot)
-
-        val treeWithSearchComponent = TreeWithSearchComponent(treeModel)
-        tree = treeWithSearchComponent.tree
-        searchField = treeWithSearchComponent.searchField
-
-        val repositoryPanel = panel(LCFlags.fill) {
-            row { searchField(growX, pushX) }
-            row { ScrollPaneFactory.createScrollPane(tree)(grow, push) }
-            row(GitlabBundle.message("clone.dialog.directory.field")) { directoryField(growX, pushX) }
+            wrapper.setContent(cloneRepositoryUI.repositoryPanel)
         }
-        repositoryPanel.border = JBEmptyBorder(UIUtil.PANEL_REGULAR_INSETS)
 
-        wrapper.setContent(repositoryPanel)
+        val gitlabHosts = gitlabDataService.gitlabData?.gitlabHosts
+        if (gitlabHosts == null && gitlabHosts?.isEmpty() == true) {
+            wrapper.setContent(tokenLoginUI.tokenLoginPanel)
+        } else {
+            wrapper.setContent(cloneRepositoryUI.repositoryPanel)
+        }
     }
 
     override fun doClone(checkoutListener: CheckoutProvider.Listener) {
-        val parent = Paths.get(directoryField.text).toAbsolutePath().parent
+        val parent = Paths.get(cloneRepositoryUI.directoryField.text).toAbsolutePath().parent
         val destinationValidation = CloneDvcsValidationUtils.createDestination("") // TODO
         if (destinationValidation != null) {
             LOG.error("Unable to create destination directory", destinationValidation.message)
@@ -97,7 +81,7 @@ class GitlabCloneDialogExtensionComponent(private val project: Project) : VcsClo
             GitlabNotifications.showError(project, GitlabBundle.message("clone.dialog.clone.failed"), GitlabBundle.message("clone.error.unable.to.find.dest"))
             return
         }
-        val directoryName = Paths.get(directoryField.text).fileName.toString()
+        val directoryName = Paths.get(cloneRepositoryUI.directoryField.text).fileName.toString()
         val parentDirectory = parent.toAbsolutePath().toString()
 
         GitCheckoutProvider.clone(project, Git.getInstance(), checkoutListener, destinationParent, "selectedUrl", directoryName, parentDirectory) // TODO
@@ -105,7 +89,7 @@ class GitlabCloneDialogExtensionComponent(private val project: Project) : VcsClo
 
     override fun doValidateAll(): List<ValidationInfo> {
         val list = ArrayList<ValidationInfo>()
-        ContainerUtil.addIfNotNull(list, CloneDvcsValidationUtils.checkDirectory(directoryField.text, directoryField.textField))
+        ContainerUtil.addIfNotNull(list, CloneDvcsValidationUtils.checkDirectory(cloneRepositoryUI.directoryField.text, cloneRepositoryUI.directoryField.textField))
         return list
     }
 
@@ -113,7 +97,6 @@ class GitlabCloneDialogExtensionComponent(private val project: Project) : VcsClo
 
     override fun onComponentSelected() {
         dialogStateListener.onOkActionNameChanged(GitlabBundle.message("clone.button"))
-        //        updateSelectedUrl() // TODO
 
         val focusManager = IdeFocusManager.getInstance(project)
         getPreferredFocusedComponent()?.let { focusManager.requestFocus(it, true) }