+++ CHANGELOG.md
@@ -9,3 +9,5 @@
 - Filter functionality into a tree
 - Added gitlab brand icon into clone dialog
 - Created basic extension point to register UI components into clone dialog to support gitlab
+- Added login dialog into a custom Gitlab installation via access token
+- store gitlab hosts and tokens to load them on restart of IntelliJ